requirements.txt
Django==1.4.3
teamcity-messages==1.6
Fabric==1.5.2
django-nose
selenium==2.29.0
pyvirtualdisplay